// packages/testing-e2e/src/codegen/sqlite/main.account-table.ts
/* File generated by Valnor. Do not edit. */
import * as valnor from "valnor";

export const Account = valnor.newSqlTable({
   name: "account",
   schema: "main",
   types: <{ Insert: IAccountInsert, Update: IAccountUpdate }>{}, 
   pk: "account_id", 
},{

   /**
    * account_id TEXT default lower(hex(randomblob(4))) || '-' || lower(hex(randomblob(2))) || '-4' || substr(lower(hex(randomblob(2))),2) || '-' || substr('89ab',abs(random()) % 4 + 1, 1) || substr(lower(hex(randomblob(2))),2) || '-' || lower(hex(randomblob(6)))
   */
   accountId: "account_id",

   /**
    * status TEXT default 'created'
   */
   status: "status",

   /**
    * email TEXT
   */
   email: "email",

   /**
    * first_name TEXT
   */
   firstName: "first_name",

   /**
    * last_name TEXT
   */
   lastName: "last_name",

   /**
    * notes TEXT
   */
   notes: "notes",

   /**
    * created_at TEXT default datetime('now')
   */
   createdAt: "created_at",

   /**
    * modified_at TEXT default datetime('now')
   */
   modifiedAt: "modified_at",
});

export type IAccountInsert = {
   accountId?: string | null;
   status?: string;
   email: string;
   firstName: string;
   lastName: string;
   notes?: string | null;
   createdAt?: string;
   modifiedAt?: string;
}

export type IAccountUpdate = Partial<IAccountInsert>;

export type IAccountSelect = {
   readonly accountId: string | null
   readonly status: string
   readonly email: string
   readonly firstName: string
   readonly lastName: string
   readonly notes: string | null
   readonly createdAt: string
   readonly modifiedAt: string
}

export type IAccountJson = valnor.JsonRow<IAccountSelect>;