// packages/testing-e2e/src/codegen/sqlite/main.order-table.ts
/* File generated by Valnor. Do not edit. */
import * as valnor from "valnor";

export const Order = valnor.newSqlTable({
   name: "order",
   schema: "main",
   types: <{ Insert: IOrderInsert, Update: IOrderUpdate }>{}, 
   pk: "order_id", 
},{

   /**
    * order_id TEXT default lower(hex(randomblob(4))) || '-' || lower(hex(randomblob(2))) || '-4' || substr(lower(hex(randomblob(2))),2) || '-' || substr('89ab',abs(random()) % 4 + 1, 1) || substr(lower(hex(randomblob(2))),2) || '-' || lower(hex(randomblob(6)))
   */
   orderId: "order_id",

   /**
    * status TEXT default 'created'
   */
   status: "status",

   /**
    * created_at TEXT
   */
   createdAt: "created_at",

   /**
    * modified_at TEXT
   */
   modifiedAt: "modified_at",

   /**
    * account_id TEXT
   */
   accountId: "account_id",
});

export type IOrderInsert = {
   orderId?: string | null;
   status?: string;
   createdAt: string;
   modifiedAt: string;
   accountId: string;
}

export type IOrderUpdate = Partial<IOrderInsert>;

export type IOrderSelect = {
   readonly orderId: string | null
   readonly status: string
   readonly createdAt: string
   readonly modifiedAt: string
   readonly accountId: string
}

export type IOrderJson = valnor.JsonRow<IOrderSelect>;